Finish

+++ API/Controllers/UsersController.cs
@@ -18,34 +18,25 @@ using Microsoft.EntityFrameworkCore;
 namespace API.Controllers
 {
     [Authorize]
-    public class UsersController : BaseApiController
+    public class UsersController(IPhotoService _photoService, IMapper _mapper, IUnitOfWork unitOfWork) : BaseApiController
     {
-        private readonly IUserRepository _userRepository;
-        private readonly IMapper _mapper;
-        private readonly IPhotoService _photoService;
-        public UsersController(IUserRepository userRepository, IMapper mapper, IPhotoService photoService)
-        {
-            _userRepository = userRepository;
-            _mapper = mapper;
-            _photoService = photoService;
-        }
         // [AllowAnonymous]
         // [Authorize(Roles = "Admin")]
         [HttpGet]
         public async Task<ActionResult<PagedList<MemberDTOs>>> GetUsers([FromQuery] UserParams userParams)
         {
-            var currentUser = await _userRepository.GetUserByUserName(User.GetUserName());
+            var currentUser = await unitOfWork.UserRepository.GetUserByUserName(User.GetUserName());
             userParams.CurrentUserName = currentUser.UserName;
             if (string.IsNullOrEmpty(userParams.Gender)){
                 userParams.Gender = currentUser.Gender == "male"? "female" : "male";
             }
-            var users = await _userRepository.GetMembersAsync(userParams);
+            var users = await unitOfWork.UserRepository.GetMembersAsync(userParams);
             Response.AddPaginationHeader(new PaginationHeader(users.CurrentPage, users.PageSize,users.TotalCount, users.TotalPages));
             return Ok(users);
         }
         // [HttpGet("{id}")]
         // public async Task<ActionResult<AppUser>> GetUser(int id){
-        //     var user = await _userRepository.GetByIdAsync(id);
+        //     var user = await unitOfWork.UserRepository.GetByIdAsync(id);
         //     if (user == null){
         //         return NotFound();
         //     }
@@ -55,28 +46,24 @@ namespace API.Controllers
         [HttpGet("{username}")]
         public async Task<ActionResult<MemberDTOs>> GetUserByUserName(string username)
         {
-            var user = await _userRepository.GetMemberAsync(username);
-            if (user == null)
-            {
-                return NotFound();
-            }
-            return Ok(user);
+            var currentUsername = User.GetUserName();
+            return await unitOfWork.UserRepository.GetMemberAsync(username, isCurrentUser : currentUsername == username);
         }
         [HttpPut]
         public async Task<ActionResult<MemberDTOs>> UpdateUser(MemberUpdateDto memberUpdateDTO)
         {
             var username = User.GetUserName();
-            var user = await _userRepository.GetUserByUserName(username);
+            var user = await unitOfWork.UserRepository.GetUserByUserName(username);
 
             if (user == null) return NotFound();
             _mapper.Map(memberUpdateDTO, user);
-            if (await _userRepository.SaveAllAsync()) return NoContent();
+            if (await unitOfWork.Complete()) return NoContent();
             return BadRequest("Failed to update user");
         }
         [HttpPost("add-photo")]
         public async Task<ActionResult<PhotoDto>> AddPhotoAsync(IFormFile file)
         {
-            var user = await _userRepository.GetUserByUserName(User.GetUserName());
+            var user = await unitOfWork.UserRepository.GetUserByUserName(User.GetUserName());
             if (user == null) return NotFound("User not found");
             var result = await _photoService.AddPhotoAsync(file);
             if (result.Error != null) return BadRequest(result.Error.Message);
@@ -85,12 +72,8 @@ namespace API.Controllers
                 Url = result.SecureUrl.AbsoluteUri,
                 PublicId = result.PublicId
             };
-            if (user.Photos.Count == 0)
-            {
-                photo.IsMain = true;
-            }
             user.Photos.Add(photo);
-            if (await _userRepository.SaveAllAsync())
+            if (await unitOfWork.Complete())
             {
                 return CreatedAtAction(nameof(GetUserByUserName),
                  new { username = user.UserName }, _mapper.Map<PhotoDto>(photo));
@@ -100,7 +83,7 @@ namespace API.Controllers
         [HttpPut("set-main-photo/{photoId}")]
         public async Task<ActionResult> SetMainPhoto(int photoId)
         {
-            var user = await _userRepository.GetUserByUserName(User.GetUserName());
+            var user = await unitOfWork.UserRepository.GetUserByUserName(User.GetUserName());
             if (user == null) return NotFound("User not found");
             var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
             if (photo == null) return NotFound("Photo not found");
@@ -108,24 +91,23 @@ namespace API.Controllers
             var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
             if (currentMain != null) currentMain.IsMain = false;
             photo.IsMain = true;
-            if (await _userRepository.SaveAllAsync()) return NoContent();
+            if (await unitOfWork.Complete()) return NoContent();
             return BadRequest("Failed to set main photo");
         }
         [HttpDelete("delete-photo/{photoId}")]
         public async Task<ActionResult> DeletePhoto(int photoId)
         {
             {
-                var user = await _userRepository.GetUserByUserName(User.GetUserName());
-                var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
-                if (photo == null) return NotFound();
-                if (photo.IsMain) return BadRequest("You cannot delete your main photo");
+                var user = await unitOfWork.UserRepository.GetUserByUserName(User.GetUserName());
+                var photo = await unitOfWork.PhotoRepository.GetPhotoById(photoId);
+                if (photo == null || photo.IsMain) return BadRequest("You cannot delete your main photo");
                 if (photo.PublicId != null)
                 {
                     var result = await _photoService.DeletePhotoAsync(photo.PublicId);
                     if (result.Error != null) return BadRequest(result.Error.Message);
                 }
                 user.Photos.Remove(photo);
-                if (await _userRepository.SaveAllAsync()) return Ok();
+                if (await unitOfWork.Complete()) return Ok();
                 return BadRequest("Failed to delete photo");
             }
         }
